set up JPA models for postgres data

+++ server/springboot/src/main/java/com/styra/tickethub_springboot/web/errors/TicketNotFoundException.java
@@ -2,9 +2,9 @@ package com.styra.tickethub_springboot.web.errors;
 
 public class TicketNotFoundException extends RuntimeException {
 
-  private Long id;
+  private Integer id;
 
-  public TicketNotFoundException(Long id) {
+  public TicketNotFoundException(Integer id) {
     super("Could not find ticket " + id);
   }
 }
